api/requirements.txt
alembic==1.7.3
aniso8601==9.0.1
appdirs==1.4.3
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.8.2
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.8
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
html5lib==1.0.1
idna==2.8
importlib-resources==5.2.2
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.1.0
lockfile==0.12.2
Mako==1.1.5
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.4.2
pyparsing==2.4.6
pyrsistent==0.18.0
python-dotenv==0.19.0
pytoml==0.1.21
pytz==2021.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.4.11
toml==0.10.2
urllib3==1.25.8
webargs==8.0.1
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.5.0
